EES-1238 - added time restrictions on when a PreRelease user can access the PreRelease page
