+++ src/GovUk.Education.ExploreEducationStatistics.Admin/Security/AuthorizationHandlers/AuthorizationHandlerUtil.cs
@@ -6,6 +6,19 @@ namespace GovUk.Education.ExploreEducationStatistics.Admin.Security.Authorizatio
 {
     public static class AuthorizationHandlerUtil
     {
+        private static readonly List<ReleaseRole> PreReleaseViewerRoles = new List<ReleaseRole>
+        {
+            ReleaseRole.PrereleaseViewer
+        };
+        
+        private static readonly List<ReleaseRole> UnrestrictedViewerRoles = new List<ReleaseRole>
+        {
+            ReleaseRole.Viewer,
+            ReleaseRole.Contributor,
+            ReleaseRole.Approver,
+            ReleaseRole.Lead
+        };
+        
         private static readonly List<ReleaseRole> EditorRoles = new List<ReleaseRole>
         {
             ReleaseRole.Contributor,
@@ -18,14 +31,29 @@ namespace GovUk.Education.ExploreEducationStatistics.Admin.Security.Authorizatio
             ReleaseRole.Approver,
         };
 
+        public static bool ContainsPreReleaseViewerRole(IEnumerable<ReleaseRole> roles)
+        {
+            return ContainsAtLeastOneCommonRole(PreReleaseViewerRoles, roles);
+        }
+
+        public static bool ContainsUnrestrictedViewerRole(IEnumerable<ReleaseRole> roles)
+        {
+            return ContainsAtLeastOneCommonRole(UnrestrictedViewerRoles, roles);
+        }
+
         public static bool ContainsEditorRole(IEnumerable<ReleaseRole> roles)
         {
-            return EditorRoles.Intersect(roles).Any();
+            return ContainsAtLeastOneCommonRole(EditorRoles, roles);
         }
 
         public static bool ContainsApproverRole(IEnumerable<ReleaseRole> roles)
         {
-            return ApproverRoles.Intersect(roles).Any();
+            return ContainsAtLeastOneCommonRole(ApproverRoles, roles);
+        }
+        
+        public static bool ContainsAtLeastOneCommonRole(IEnumerable<ReleaseRole> roles1, IEnumerable<ReleaseRole> roles2)
+        {
+            return roles1.Intersect(roles2).Any();
         }
     }
 }
