+++ src/GovUk.Education.ExploreEducationStatistics.Admin/Services/Interfaces/IPreReleaseService.cs
@@ -3,18 +3,21 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 using GovUk.Education.ExploreEducationStatistics.Admin.Models.Api;
 using GovUk.Education.ExploreEducationStatistics.Common.Model;
+using GovUk.Education.ExploreEducationStatistics.Content.Model;
 using Microsoft.AspNetCore.Mvc;
 
 namespace GovUk.Education.ExploreEducationStatistics.Admin.Services.Interfaces
 {
+    public enum PreReleaseWindowStatus
+    {
+        NoneSet,
+        Before,
+        Within,
+        After
+    } 
+    
     public interface IPreReleaseService
     {
-        Task<Either<ActionResult, List<PrereleaseCandidateViewModel>>> GetAvailablePreReleaseContactsAsync();
-        
-        Task<Either<ActionResult, List<PrereleaseCandidateViewModel>>> GetPreReleaseContactsForReleaseAsync(Guid releaseId);
-
-        Task<Either<ActionResult, List<PrereleaseCandidateViewModel>>> AddPreReleaseContactToReleaseAsync(Guid releaseId, string email);
-
-        Task<Either<ActionResult, List<PrereleaseCandidateViewModel>>> RemovePreReleaseContactFromReleaseAsync(Guid releaseId, string email);
+        PreReleaseWindowStatus GetPreReleaseWindowStatus(Release release, DateTime referenceTime);
     }
 }
