+++ src/GovUk.Education.ExploreEducationStatistics.Admin/Startup.cs
@@ -41,7 +41,6 @@ using Microsoft.AspNetCore.SpaServices.ReactDevelopmentServer;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.FileProviders;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using Microsoft.IdentityModel.Tokens;
@@ -169,6 +168,8 @@ namespace GovUk.Education.ExploreEducationStatistics.Admin
             {
                 options.ClaimsIdentity.UserIdClaimType = ClaimTypes.NameIdentifier;
             });
+            
+            services.Configure<PreReleaseOptions>(Configuration);
 
             services.AddAuthorization(options =>
             {
@@ -290,6 +291,7 @@ namespace GovUk.Education.ExploreEducationStatistics.Admin
             services.AddTransient<IReleaseRepository, ReleaseRepository>();
             services.AddTransient<IMethodologyService, MethodologyService>();
             services.AddTransient<IDataBlockService, DataBlockService>();
+            services.AddTransient<IPreReleaseContactsService, PreReleaseContactsService>();
             services.AddTransient<IPreReleaseService, PreReleaseService>();
 
             services.AddTransient<IManageContentPageService, ManageContentPageService>();
