+++ src/GovUk.Education.ExploreEducationStatistics.Admin/appsettings.Development.json
@@ -20,5 +20,11 @@
     "Key": {
       "Type": "Development"
     }
+  },
+  "PreReleaseAccess": {
+    "AccessWindow": {
+      "MinutesBeforeReleaseTimeStart": 1440,
+      "MinutesBeforeReleaseTimeEnd": -1440
+    }
   }
 }
