+++ src/GovUk.Education.ExploreEducationStatistics.Admin/appsettings.json
@@ -26,5 +26,11 @@
     "ResponseType": "code id_token",
     "TokenValidationParameters.NameClaimType": "name",
     "TokenValidationParameters.RoleClaimType": "role"
+  },
+  "PreReleaseAccess": {
+    "AccessWindow": {
+      "MinutesBeforeReleaseTimeStart": 870,
+      "MinutesBeforeReleaseTimeEnd": 0
+    }
   }
 }
